+++ packages/server/src/services/hermes/hermes-cli.ts
