+++ packages/server/src/services/hermes/hermes.ts
@@ -1,4 +1,4 @@
-import { config } from '../config'
+import { config } from '../../config'
 
 const UPSTREAM = config.upstream.replace(/\/$/, '')
